Remove unneeded castings

+++ src/deluge/model/instrument/kit.cpp
@@ -672,14 +672,14 @@ void Kit::renderOutput(ModelStack* modelStack, StereoSample* outputBuffer, Stere
 			NonAudioDrum* nonAudioDrum = (NonAudioDrum*)thisNoteRow->drum;
 
 			uint32_t gateThreshold = (uint32_t)nonAudioDrum->arpeggiatorGate + 2147483648;
-			uint32_t noteProbability = (uint32_t)nonAudioDrum->arpeggiatorNoteProbability;
-			uint32_t ratchetProbability = (uint32_t)nonAudioDrum->arpeggiatorRatchetProbability;
-			uint32_t ratchetAmount = (uint32_t)nonAudioDrum->arpeggiatorRatchetAmount;
-			uint32_t sequenceLength = (uint32_t)nonAudioDrum->arpeggiatorSequenceLength;
-			uint32_t rhythm = (uint32_t)nonAudioDrum->arpeggiatorRhythm;
-			uint32_t spreadVelocity = (uint32_t)nonAudioDrum->arpeggiatorSpreadVelocity;
-			uint32_t spreadGate = (uint32_t)nonAudioDrum->arpeggiatorSpreadGate;
-			uint32_t spreadOctave = (uint32_t)nonAudioDrum->arpeggiatorSpreadOctave;
+			uint32_t noteProbability = nonAudioDrum->arpeggiatorNoteProbability;
+			uint32_t ratchetProbability = nonAudioDrum->arpeggiatorRatchetProbability;
+			uint32_t ratchetAmount = nonAudioDrum->arpeggiatorRatchetAmount;
+			uint32_t sequenceLength = nonAudioDrum->arpeggiatorSequenceLength;
+			uint32_t rhythm = nonAudioDrum->arpeggiatorRhythm;
+			uint32_t spreadVelocity = nonAudioDrum->arpeggiatorSpreadVelocity;
+			uint32_t spreadGate = nonAudioDrum->arpeggiatorSpreadGate;
+			uint32_t spreadOctave = nonAudioDrum->arpeggiatorSpreadOctave;
 
 			uint32_t phaseIncrement = nonAudioDrum->arpSettings.getPhaseIncrement(
 			    getFinalParameterValueExp(paramNeutralValues[deluge::modulation::params::GLOBAL_ARP_RATE],
@@ -1009,14 +1009,14 @@ int32_t Kit::doTickForwardForArp(ModelStack* modelStack, int32_t currentPos) {
 			if (thisNoteRow->drum->type == DrumType::MIDI || thisNoteRow->drum->type == DrumType::GATE) {
 				NonAudioDrum* nonAudioDrum = (NonAudioDrum*)thisNoteRow->drum;
 
-				uint32_t sequenceLength = (uint32_t)nonAudioDrum->arpeggiatorSequenceLength;
-				uint32_t rhythm = (uint32_t)nonAudioDrum->arpeggiatorRhythm;
-				uint32_t noteProbability = (uint32_t)nonAudioDrum->arpeggiatorNoteProbability;
-				uint32_t ratchetAmount = (uint32_t)nonAudioDrum->arpeggiatorRatchetAmount;
-				uint32_t ratchetProbability = (uint32_t)nonAudioDrum->arpeggiatorRatchetProbability;
-				uint32_t spreadVelocity = (uint32_t)nonAudioDrum->arpeggiatorSpreadVelocity;
-				uint32_t spreadGate = (uint32_t)nonAudioDrum->arpeggiatorSpreadGate;
-				uint32_t spreadOctave = (uint32_t)nonAudioDrum->arpeggiatorSpreadOctave;
+				uint32_t sequenceLength = nonAudioDrum->arpeggiatorSequenceLength;
+				uint32_t rhythm = nonAudioDrum->arpeggiatorRhythm;
+				uint32_t noteProbability = nonAudioDrum->arpeggiatorNoteProbability;
+				uint32_t ratchetAmount = nonAudioDrum->arpeggiatorRatchetAmount;
+				uint32_t ratchetProbability = nonAudioDrum->arpeggiatorRatchetProbability;
+				uint32_t spreadVelocity = nonAudioDrum->arpeggiatorSpreadVelocity;
+				uint32_t spreadGate = nonAudioDrum->arpeggiatorSpreadGate;
+				uint32_t spreadOctave = nonAudioDrum->arpeggiatorSpreadOctave;
 
 				drum->arpeggiator.updateParams(sequenceLength, rhythm, noteProbability, ratchetAmount,
 				                               ratchetProbability, spreadVelocity, spreadGate, spreadOctave);
